Replaced Int64 with Int as StorableValue
Simplified use on iOS, tvOS, and macOS. The library assumes a 64 bit Int, and will therefore not support current generations of the SX chips in Apple Watch

+++ SwiftyDB/Classes/Database/SQLite/Utilities/SQLiteQueryFactory.swift
@@ -48,7 +48,7 @@ extension SQLiteFilterDeparameterizer {
             return "'\(string)'"
         case let double as Double:
             return "\(double)"
-        case let int as Int64:
+        case let int as Int:
             return "\(int)"
         default:
             return "NULL"
